+++ main/models.py
@@ -30,6 +30,9 @@ class Lapangan(models.Model):
     review = models.TextField(blank=True, null=True)
     peraturan = models.TextField(blank=True, null=True)
     fasilitas = models.TextField(blank=True, null=True)
+
+    rating = models.FloatField(default=0.0)  
+    original_rating = models.FloatField(default=0.0)  
     
     def __str__(self):
         return self.nama
@@ -84,18 +87,3 @@ class Event(models.Model):
     def __str__(self):
         return self.nama
  
-
-class Review(models.Model):
-    lapangan = models.ForeignKey(Lapangan, on_delete=models.CASCADE, related_name='reviews')
-    reviewer_name = models.CharField(max_length=255, default="Anonim")
-    rating = models.DecimalField(max_digits=2, decimal_places=1, default=0.0) # 0.0 - 5.0
-    review_text = models.TextField()
-    tanggal_dibuat = models.DateTimeField(auto_now_add=True)
-    gambar = models.URLField(blank=True, null=True) 
-    session_key = models.CharField(max_length=255, blank=True, null=True)
-    
-    class Meta:
-        ordering = ['-tanggal_dibuat']
-
-    def __str__(self):
-        return f'Review oleh {self.reviewer_name} untuk {self.lapangan.nama} - Rating: {self.rating}'
