+++ main/views.py
@@ -3,8 +3,8 @@ from django.http import JsonResponse
 from django.core import serializers
 
 from main import models
-from main.forms import LapanganForm, CoachForm, EventForm, ReviewForm
-from main.models import Lapangan, Coach, Event, Review
+from main.forms import LapanganForm, CoachForm, EventForm
+from main.models import Lapangan, Coach, Event
 from django.db.models import Q
 
 from django.contrib import messages
@@ -22,7 +22,7 @@ from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_POST
 from django.utils.html import strip_tags;
 
-from django.db.models import Q, Avg
+
 
 def show_main(request):
     context = {
@@ -278,200 +278,3 @@ def show_lapangan_dashboard(request):
     }
     return render(request, 'lapangan/dashboard_lapangan.html', context)
 
-#==========================================================
-# Semua fungsi buat review hehe
-#==========================================================
-
-def show_review_lapangan(request, lapangan_id):
-    try:
-        lapangan = Lapangan.objects.get(pk=lapangan_id)
-    except Lapangan.DoesNotExist:
-        return HttpResponse("Lapangan tidak ditemukan.", status=404)
-        
-    reviews = Review.objects.filter(lapangan=lapangan)
-    review_form = ReviewForm()
-
-    context = {
-        'lapangan': lapangan,
-        'reviews': reviews,
-        'review_form': review_form,
-        'is_authenticated': request.user.is_authenticated 
-    }
-    return render(request, 'lapangan/review_lapangan.html', context)
-
-def show_feeds_review_lapangan(request, lapangan_id):
-    try:
-        lapangan = Lapangan.objects.get(pk=lapangan_id)
-    except Lapangan.DoesNotExist:
-        return HttpResponse("Lapangan tidak ditemukan.", status=404)
-        
-    context = {
-        'lapangan': lapangan,
-    }
-    return render(request, 'lapangan/feeds_review_lapangan.html', context)
-
-def show_form_review_lapangan(request, lapangan_id):
-    try:
-        lapangan = Lapangan.objects.get(pk=lapangan_id)
-    except Lapangan.DoesNotExist:
-        return HttpResponse("Lapangan tidak ditemukan.", status=404)
-
-    if request.method == 'POST':
-        form = ReviewForm(request.POST)
-        if form.is_valid():
-            new_review = form.save(commit=False)
-            new_review.lapangan = lapangan
-            new_review.save()
-            
-            average_rating = Review.objects.filter(lapangan=lapangan).aggregate(Avg('rating'))['rating__avg'] or 0.0
-            lapangan.rating = round(average_rating, 1)
-            lapangan.save()
-            
-            return redirect('main:show_feeds_review_lapangan', lapangan_id=lapangan.id)
-
-    else:
-        form = ReviewForm()
-
-    context = {
-        'lapangan': lapangan,
-        'review_form': form,
-    }
-    return render(request, 'lapangan/form_review_lapangan.html', context)
-
-
-@csrf_exempt
-def add_review_lapangan(request, lapangan_id):
-    if request.method == 'POST':
-        try:
-            lapangan = Lapangan.objects.get(pk=lapangan_id)
-        except Lapangan.DoesNotExist:
-            return JsonResponse({'status': 'error', 'message': 'Lapangan tidak ditemukan.'}, status=404)
-            
-        form = ReviewForm(request.POST)
-        if form.is_valid():
-            new_review = form.save(commit=False)
-            new_review.lapangan = lapangan
-            new_review.save()
-            
-            average_rating = Review.objects.filter(lapangan=lapangan).aggregate(Avg('rating'))['rating__avg'] or 0.0
-            lapangan.rating = round(average_rating, 1)
-            lapangan.save()
-
-            return JsonResponse({'status': 'success', 'message': 'Review berhasil ditambahkan!'}, status=201)
-        else:
-            errors = dict(form.errors.items())
-            return JsonResponse({'status': 'error', 'errors': errors}, status=400)
-    
-    return JsonResponse({'status': 'error', 'message': 'Metode permintaan tidak valid.'}, status=405)
-
-    
-@csrf_exempt
-def update_review(request, review_id):
-    try:
-        review = Review.objects.get(pk=review_id)
-        if review.session_key != request.session.session_key:
-            return JsonResponse({'status': 'error', 'message': 'Anda tidak memiliki izin untuk mengedit review ini.'}, status=403)
-            
-        if request.method == 'POST':
-            form = ReviewForm(request.POST, instance=review)
-            if form.is_valid():
-                form.save()
-                
-                lapangan = review.lapangan
-                average_rating = Review.objects.filter(lapangan=lapangan).aggregate(Avg('rating'))['rating__avg'] or 0.0
-                lapangan.rating = round(average_rating, 1)
-                lapangan.save()
-                
-                return JsonResponse({'status': 'success', 'message': 'Review berhasil diupdate!'})
-            else:
-                errors = dict(form.errors.items())
-                return JsonResponse({'status': 'error', 'errors': errors}, status=400)
-                
-    except Review.DoesNotExist:
-        return JsonResponse({'status': 'error', 'message': 'Review tidak ditemukan.'}, status=404)
-
-@csrf_exempt
-@require_POST
-def delete_review(request, review_id):
-    try:
-        review = Review.objects.get(pk=review_id)
-        lapangan = review.lapangan
-        
-        review.delete()
-        
-        average_rating = Review.objects.filter(lapangan=lapangan).aggregate(Avg('rating'))['rating__avg'] or 0.0
-        lapangan.rating = round(average_rating, 1)
-        lapangan.save()
-        
-        return JsonResponse({'status': 'success', 'message': 'Review berhasil dihapus.'}, status=200)
-    except Review.DoesNotExist:
-        return JsonResponse({'status': 'error', 'message': 'Review tidak ditemukan.'}, status=404)
-    
-
-def get_reviews_json(request, lapangan_id):
-    try:
-        lapangan = Lapangan.objects.get(pk=lapangan_id)
-    except Lapangan.DoesNotExist:
-        return JsonResponse({'detail': 'Lapangan not found'}, status=404)
-        
-    reviews = Review.objects.filter(lapangan=lapangan).order_by('-tanggal_dibuat')
-    
-    # Untuk sementara, set semua bisa edit/hapus (nanti ganti dengan session logic)
-    current_session = request.session.session_key
-    can_delete_review = True  # Ganti dengan logic session nanti
-
-    data = [
-        {
-            'id': review.id,
-            'reviewer_name': review.reviewer_name,
-            'rating': float(review.rating),
-            'review_text': review.review_text,
-            'tanggal_dibuat': review.tanggal_dibuat.strftime("%Y-%m-%d %H:%M"),
-            'gambar': review.gambar,
-            'can_edit': True, 
-            'can_delete': True  
-        }
-        for review in reviews
-    ]
-    return JsonResponse(data, safe=False)
-
-def get_single_review_json(request, review_id):
-    try:
-        review = Review.objects.get(pk=review_id)
-        data = {
-            'id': review.id,
-            'reviewer_name': review.reviewer_name,
-            'rating': float(review.rating),
-            'review_text': review.review_text,
-            'gambar': review.gambar,
-        }
-        return JsonResponse(data)
-    except Review.DoesNotExist:
-        return JsonResponse({'detail': 'Review not found'}, status=404)
-
-def show_edit_review_lapangan(request, review_id):
-    try:
-        review = Review.objects.get(pk=review_id)
-            
-        if request.method == 'POST':
-            form = ReviewForm(request.POST, instance=review)
-            if form.is_valid():
-                form.save()
-                
-                lapangan = review.lapangan
-                average_rating = Review.objects.filter(lapangan=lapangan).aggregate(Avg('rating'))['rating__avg'] or 0.0
-                lapangan.rating = round(average_rating, 1)
-                lapangan.save()
-                
-                return redirect('main:show_feeds_review_lapangan', lapangan_id=review.lapangan.id)
-        else:
-            form = ReviewForm(instance=review)
-
-        context = {
-            'review': review,
-            'review_form': form,
-        }
-        return render(request, 'lapangan/edit_review_lapangan.html', context)
-        
-    except Review.DoesNotExist:
-        return HttpResponse("Review tidak ditemukan.", status=404)
